Enabled keyboard controls.  moved screenshot to it's own component.  Ready for review.
