+++ client/src/router/index.js
@@ -1,7 +1,5 @@
 import Vue from 'vue'
 import Router from 'vue-router'
-import VideoPlayer from '@/pages/VideoPlayer'
-
 Vue.use(Router)
 
 export default new Router({
@@ -9,7 +7,6 @@ export default new Router({
   routes: [
     {
       path: '/?filter=:filter',
-      component: VideoPlayer
     }
   ]
 })
